refactor(session): use broker and hunk adapter boundaries
